Replace parcel with webpack
