+++ .gitignore
@@ -1,4 +1,4 @@
-.DS_Store
-.cache
-dist
-node_modules
+**/.DS_Store
+/node_modules
+/dist
+/.cache
