--- README.md
@@ -1,61 +1,0 @@
-# Travel journal
-
-Travel journal is a personal travel tracking webpage. 
-
-
-## Project objectives 
-
-This is the project I am working on as a part of the JavaScript learning process. 
-
-Some of the objectives include: 
-
-:ballot_box_with_check: learn the JS basics
-
-:ballot_box_with_check: use parcel and npm to set up the project 
-
-:ballot_box_with_check: dinamically rendered HTML  
-
-:ballot_box_with_check: learn the concept of functional programming
-
-:ballot_box_with_check: JS modules 
-
-:ballot_box_with_check: learn and implement code design patterns
-
-:ballot_box_with_check: refactoring
-
-:ballot_box_with_check: use local storage 
-
-:ballot_box_with_check: apply BEM principles 
-
-:ballot_box_with_check: use a serverless database
-
-:black_square_button:   style the page 
-
-
-
-## Installation
-
-Install dependencies
-
-```
-npm install
-```
-
-Run the project 
-
-```
-npm run dev
-```
-
-## Current stage
-In development
-
-## Usage
-
-
-## Contributing
-Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
-
-## Acknowledgments 
-Inspiration and guidence
-@sandrina-p
