--- src/main.js
@@ -1,6 +1,0 @@
-import destinationsList from "./scripts/store.js";
-import moduleDestinations from "./scripts/modules/list.js";
-import addNewCity from "./scripts/modules/add_destination.js";
-import events from "./scripts/modules/pubsub.js";
-moduleDestinations.init();
-addNewCity.init();
