+++ src/styles/_main-page.scss
@@ -1,0 +1,329 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+//Basic styles
+html {
+  background: $grey;
+  font-size: 10px;
+  font-family: $main-font;
+}
+
+li {
+  list-style: none;
+}
+
+a {
+  text-decoration: none;
+}
+
+button {
+  outline: none;
+  border: none;
+  cursor: pointer;
+}
+
+.container {
+  width: 90vw;
+  margin: auto;
+}
+
+.container--space-between {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+//Navbar
+.navbar {
+  height: 7rem;
+  width: 100%;
+  background: transparent;
+  letter-spacing: 2px;
+  font-size: 2rem;
+  line-height: 1.2;
+  transition: all 0.4s ease;
+  font-weight: normal;
+  font-style: normal;
+  position: absolute;
+  top: 0;
+  left: 0;
+  &__logo {
+    font-family: $cursive;
+    color: white;
+  }
+  &__links {
+    margin: 4rem;
+    display: flex;
+  }
+  &__link {
+    margin-left: 4rem;
+    color: white;
+  }
+}
+
+//Header
+.header {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 15vh;
+  &__img {
+    height: 100vh;
+    width: 100vw;
+    object-fit: cover;
+    min-width: 50rem;
+    display: block;
+    margin: auto;
+    &:hover {
+      opacity: 100;
+    }
+  }
+}
+
+//Gallery display buttons (visited, bucketlist, all)
+.display__buttons {
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  padding-bottom: 3rem;
+  height: 5rem;
+  font-size: 1.8rem;
+  .display__button {
+    display: inline-block;
+    margin: 2rem;
+    font-family: $main-font;
+    font-size: 1em;
+    font-weight: 100;
+    text-transform: uppercase;
+    opacity: 1;
+    //Clicked button
+    &--active {
+      border-bottom: 1px solid black;
+      &:first-child,
+      &:last-child {
+        border-bottom: none;
+      }
+    }
+  }
+}
+
+.add-city-form {
+  &--closed {
+    visibility: hidden;
+    height: 0;
+    width: 0;
+  }
+  &--open {
+    visibility: visible;
+    background: darken($color: $grey, $amount: 5%);
+    z-index: 1000;
+    font-size: 1.9rem;
+    position: fixed;
+    top: 20vh;
+    bottom: 20vh;
+    left: 30vw;
+    right: 30vw;
+    height: auto;
+    width: auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    .form__input {
+      border: none;
+      padding: 0.5rem;
+    }
+    .form__label,
+    .form__radio-buttons,
+    .form__submit {
+      margin-top: 2.5rem;
+      margin-bottom: 0.5rem;
+    }
+    .form__submit {
+      text-transform: uppercase;
+      font-weight: 500;
+      padding: 1rem 2.5rem;
+      font-size: 1.5rem;
+      background: darken($color: $grey, $amount: 10%);
+    }
+    .form__close {
+      background: darken($color: $grey, $amount: 5%);
+      &::before {
+        content: "\f100";
+        display: block;
+        position: relative;
+        left: 20rem;
+        bottom: 38rem;
+      }
+    }
+  }
+}
+
+//Gallery
+.gallery {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  //fixes the last row formating
+  &::after {
+    content: "";
+    width: 50rem;
+  }
+}
+
+.gallery-img--container {
+  width: 50rem;
+}
+.gallery-img {
+  width: 100%;
+  height: 30rem;
+  display: block;
+  opacity: 1;
+  object-fit: cover;
+  transition: opacity 0.3s ease;
+  &:hover {
+    opacity: 0.8;
+  }
+}
+
+.gallery-delete-btn {
+  position: relative;
+  top: -28rem;
+  right: 2rem;
+  float: right;
+  opacity: 0;
+
+  //delete icon
+  &.flaticon-001-cancel-3 {
+    background: transparentize($color: white, $amount: 0.5);
+    border-radius: 50%;
+    width: 4rem;
+    height: 4rem;
+    &::before {
+      margin: 0;
+    }
+  }
+}
+
+.gallery-img:hover ~ .gallery-delete-btn {
+  opacity: 1;
+}
+
+.gallery-delete-btn:hover {
+  opacity: 1;
+}
+
+//Switching lists on click
+.gallery figure.visited,
+.gallery figure.bucketlist {
+  display: block;
+}
+
+.gallery.visited figure.visited {
+  display: block;
+}
+
+.gallery.visited figure.bucketlist {
+  display: none;
+}
+
+.gallery.bucketlist figure.bucketlist {
+  display: block;
+}
+
+.gallery.bucketlist figure.visited {
+  display: none;
+}
+
+//img captions
+.caption {
+  line-height: 1.5;
+  margin: 1rem 0;
+  display: flex;
+  .caption-text {
+    margin-left: 5px;
+  }
+  .caption__city {
+    font-weight: 300;
+    font-size: 2rem;
+    text-transform: capitalize;
+    letter-spacing: 1px;
+  }
+  .caption__country {
+    font-weight: 300;
+    font-size: 1.2rem;
+    letter-spacing: 2px;
+    text-transform: uppercase;
+  }
+}
+
+.gallery-checkbox {
+  position: relative;
+  top: 0.8rem;
+  width: 1.5rem;
+  height: 1.5rem;
+  color: white;
+  border: 1px solid rgb(163, 161, 161);
+  border-radius: 4px;
+  appearance: none;
+  outline: 0;
+  cursor: pointer;
+  transition: background 175ms cubic-bezier(0.1, 0.1, 0.25, 1);
+  &::before {
+    position: absolute;
+    content: "";
+    display: block;
+    top: -3px;
+    left: 2px;
+    width: 7px;
+    height: 13px;
+    border-style: solid;
+    border-color: black;
+    border-width: 0 2px 2px 0;
+    transform: rotate(45deg);
+    opacity: 0;
+  }
+  &:checked {
+    background: $grey;
+    border: $grey;
+    &::before {
+      opacity: 1;
+    }
+  }
+}
+//cross out visited city
+.gallery.visited figure.visited .caption__city,
+.gallery figure.visited .caption__city {
+  text-decoration: line-through;
+  text-decoration-thickness: 0.7px;
+  transition: clip-path 200ms cubic-bezier(0.25, 0.46, 0.45, 0.94);
+}
+
+.caption__city {
+  font-weight: 300;
+  font-size: 2rem;
+  text-transform: capitalize;
+  letter-spacing: 1px;
+  position: relative;
+  cursor: pointer;
+  user-select: none;
+  transition: clip-path 200ms cubic-bezier(0.25, 0.46, 0.45, 0.94);
+}
+
+.gallery
+  .bucketlist
+  .caption
+  .gallery-checkbox:checked
+  > .caption-text
+  > .caption__city {
+  &::before {
+    position: absolute;
+    content: attr(data-content);
+    color: lightgray;
+    text-decoration: line-through;
+    text-decoration-thickness: 3px;
+    text-decoration-color: black;
+  }
+}
