+++ src/styles/_variables.scss
@@ -1,0 +1,4 @@
+$grey: #f1f1f1;
+$text: black;
+$cursive: "bimboballpoint", sans-serif;
+$main-font: "TwCenW01-Medium", sans-serif;
